chore: separate note and schema graph logic
