+++ packages/common-all/src/types/typesv2.ts
@@ -521,7 +521,7 @@ export type GraphViewMessage = DMessage<GraphViewMessageType, { id: string }>;
 export enum DendronWebViewKey {
   CONFIGURE = "dendron.configure",
   NOTE_GRAPH = "dendron.graph",
-  SCHEMA_GRAPH = "dendron.graph",
+  SCHEMA_GRAPH = "dendron.graph-schema",
 }
 
 export enum DendronTreeViewKey {
